moved test, added coverage, added badges, cached node_modules in pipelines
