+++ .github/workflows/ci.yaml
@@ -15,8 +15,8 @@ jobs:
     steps:
       - uses: actions/checkout@v2
       - uses: actions/setup-node@v2
-      - name: install dependencies
-        run: npm i
+      - name: cache node modules
+        uses: ./.github/actions/cache-node-modules
       - name: check formatting
         run: npm run format-check
   continuous-integration:
@@ -25,7 +25,7 @@ jobs:
     steps:
       - uses: actions/checkout@v2
       - uses: actions/setup-node@v2
-      - name: install dependencies
-        run: npm i
+      - name: cache node modules
+        uses: ./.github/actions/cache-node-modules
       - name: test
-        run: npm test
+        run: npm test:ci
