+++ README.md
@@ -1,3 +1,15 @@
+![Continuous Deployment](https://github.com/noahvarghese/capstone-client/actions/workflows/cd.yaml/badge.svg)
+<br />
+![Continuous Integration](https://github.com/noahvarghese/capstone-client/actions/workflows/ci.yaml/badge.svg)
+<br />
+![Statements](https://img.shields.io/badge/statements-77.38%25-red.svg?style=flat)
+<br />
+![Branches](https://img.shields.io/badge/branches-65.82%25-red.svg?style=flat)
+<br />
+![Functions](https://img.shields.io/badge/functions-74.8%25-red.svg?style=flat)
+<br />
+![Lines](https://img.shields.io/badge/lines-78.26%25-red.svg?style=flat)
+
 # OnBoard - Frontend
 
 ## About
@@ -18,12 +30,12 @@ They may be moved into their own repo at the end of this to showcase all parts o
 
 ## Areas of Focus
 
-- React
-- Redux
-- TypeScript
-- CI/CD
-- AWS (S3, ACM, CloudFront, Route53, CodePipeline, CodeBuild)
+-   React
+-   Redux
+-   TypeScript
+-   CI/CD
+-   AWS (S3, ACM, CloudFront, Route53, CodePipeline, CodeBuild)
 
 ## Environment
 
-- Just rely on NODE_ENV to dictate what to use as any environment variables used will be visible client side
+-   Just rely on NODE_ENV to dictate what to use as any environment variables used will be visible client side
