+++ __test__/attributes/InviteAttributes.ts
