+++ package.json
@@ -22,8 +22,10 @@
     },
     "scripts": {
         "start": "react-scripts start",
+        "badges": "istanbul-badges-readme --coverageDir='./coverage'",
         "build": "react-scripts build",
-        "test": "react-scripts test",
+        "test": "react-scripts test --env=jsdom --coverage",
+        "test:ci": "CI=true npm test",
         "eject": "react-scripts eject",
         "format": "prettier --write **/*.ts",
         "format-check": "prettier --check **/*.ts"
@@ -61,8 +63,24 @@
         "@types/react-router-dom": "^5.1.8",
         "@types/react-table": "^7.7.7",
         "@types/validator": "^13.6.3",
+        "istanbul-badges-readme": "^1.6.0",
         "prettier": "^2.4.1",
         "typescript": "^4.2.4"
     },
-    "jest": {}
+    "jest": {
+        "collectCoverageFrom": [
+            "**/*.tsx"
+        ],
+        "coveragePathIgnorePatterns": [
+            "node_modules",
+            "test.tsx",
+            "__test__"
+        ],
+        "coverageReporters": [
+            "json-summary",
+            "text",
+            "lcov",
+            "clover"
+        ]
+    }
 }
