+++ src/App.test.tsx
@@ -1,4 +1,4 @@
-import { render, screen, cleanup, act } from "../test/test-utils";
+import { render, screen, cleanup, act } from "../__test__/test-utils";
 import App from "./App";
 import { createMemoryHistory, MemoryHistory } from "history";
