+++ src/setupTests.ts
@@ -3,3 +3,16 @@
 // expect(element).toHaveTextContent(/react/i)
 // learn more: https://github.com/testing-library/jest-dom
 import "@testing-library/jest-dom";
+import dotenv from "dotenv";
+
+dotenv.config();
+
+const { TIMEOUT_MULTIPLIER } = process.env;
+
+const DEFAULT_MULTIPLIER = 1;
+
+const multiplier = isNaN(Number(TIMEOUT_MULTIPLIER))
+    ? Number(TIMEOUT_MULTIPLIER)
+    : DEFAULT_MULTIPLIER;
+
+jest.setTimeout(10000 * multiplier);
